Move container gateway tests to its feature dir
